added analysis page and some other general ui enhancements

+++ BikesModel/Model/Bank.cs
@@ -155,7 +155,7 @@ namespace Bikes.Model.Banking
             return accounts;
         }
 
-        public static Payment deposit(Rider rider, double amount, string description)
+        public static Payment deposit(Rider rider, decimal amount, string description)
         {
             Payment payment = new Payment();
